Modify the way designate the socket id. Fix, again, keep center aliver. Remove some useless stuffs. Remove isConnected/isConnecting/isConnectionFailed, use SocketStatus instead. Combine some functions(Connection relelated) to SocketBase. Combine some functions(Set IP/Port) to WvsBase. Rename some variables LEL.

+++ WvsCenter/WvsCenter.cpp
@@ -46,7 +46,7 @@ LocalServerEntry * WvsCenter::GetChannel(int idx)
 
 int WvsCenter::GetChannelCount()
 {
-	return m_mChannel.size();
+	return (int)m_mChannel.size();
 }
 
 void WvsCenter::Init()
